fix compile and import
