+++ python/pyfury/__init__.py
@@ -27,20 +27,16 @@ try:
 except ImportError:
     ENABLE_FURY_CYTHON_SERIALIZATION = False
 
+from pyfury._fury import (  # noqa: F401,F403,F811 # pylint: disable=unused-import
+    Fury,
+    Language,
+    OpaqueObject,
+)
 if ENABLE_FURY_CYTHON_SERIALIZATION:
     from pyfury._serialization import (  # noqa: F401,F811
         Fury,
-        Language,
-        ClassInfo,
-        OpaqueObject,
     )
-else:
-    from pyfury._fury import (  # noqa: F401,F403,F811 # pylint: disable=unused-import
-        Fury,
-        Language,
-        ClassInfo,
-        OpaqueObject,
-    )
+
 from pyfury._struct import (  # noqa: F401,F403,F811 # pylint: disable=unused-import
     ComplexObjectSerializer,
 )
